Cast download progress counters to size_t to avoid data loss and wrong progress report
Before:
[DEBUG ] : SWUPDATE running :  [channel_callback_xferinfo] : Downloaded 1% (1 of 880 kB).
[DEBUG ] : SWUPDATE running :  [channel_callback_xferinfo] : Downloaded 2% (1 of 1744 kB).
[DEBUG ] : SWUPDATE running :  [channel_callback_xferinfo] : Downloaded 3% (1 of 2608 kB).

After
[DEBUG ] : SWUPDATE running :  [channel_callback_xferinfo] : Downloaded 1% (880 of 86760 kB)
[DEBUG ] : SWUPDATE running :  [channel_callback_xferinfo] : Downloaded 2% (1744 of 86760 kB)
[DEBUG ] : SWUPDATE running :  [channel_callback_xferinfo] : Downloaded 3% (2608 of 86760 kB)

+++ corelib/channel_curl.c
@@ -439,7 +439,9 @@ static int channel_callback_xferinfo(void *p, curl_off_t dltotal, curl_off_t dln
 	else
 		data->percent = percent;
 
-	DEBUG("Downloaded %d%% (%lu of %lu kB).", percent, dlnow / 1024, dltotal / 1024);
+	DEBUG("Downloaded %d%% (%zu of %zu kB).", percent,
+		(size_t)dlnow / 1024,
+		(size_t)dltotal / 1024);
 	swupdate_download_update(percent, dltotal);
 
 	return 0;
